+++ src/web/assets/src/js/typesense.ts
@@ -1,6 +1,0 @@
-// Import our CSS
-import '@/css/app.pcss'
-
-const main = async () => {};
-
-main().then( (root) => {});